token base auth

+++ app/Controllers/OrdersController.php
@@ -10,17 +10,25 @@ class OrdersController extends Controller
 {
     public function index()
     {
+        $token = Request::Token();
+        if (!$token || !$this->checkToken($token)) return $this->unauthorized();
+
         return Response::json(Order::get());
     }
 
     public function show(int $id)
     {
+        $token = Request::Token();
+        if (!$token || !$this->checkToken($token)) return $this->unauthorized();
+
         return Response::json(Order::get($id));
     }
 
     public function store()
     {
         $data = Request::POST();
+        $token = Request::Token();
+        if (!$token || !$this->checkToken($token)) return $this->unauthorized();
 
         return Response::json(Order::create(array_keys($data), array_values($data)));
     }
@@ -28,12 +36,17 @@ class OrdersController extends Controller
     public function update(int $id)
     {
         $data = Request::PUT();
+        $token = Request::Token();
+        if (!$token || !$this->checkToken($token)) return $this->unauthorized();
 
         return Response::json(Order::update($id, array_keys($data), array_values($data)));
     }
 
     public function destroy(int $id)
     {
+        $token = Request::Token();
+        if (!$token || !$this->checkToken($token)) return $this->unauthorized();
+
         if (Order::delete($id)) return Response::json(["detail" => "order deleted successfuly"]);
         Response::statusCode(404);
         return Response::json(["detail" => "order not found."]);
